+++ mor/apps/CMakeLists.txt
@@ -4,16 +4,17 @@ feelpp_add_application( onlinerun SRCS crbonlinerun.cpp LINK_LIBRARIES  Feelpp::
 add_dependencies(feelpp_mor_apps feelpp_mor_onlinerun)
 if ( TARGET Feelpp::feelpp_toolbox_heat_lib )
 
-  mor_add_library( toolboxmor_plugin_heat
-    SRCS toolboxmor_heat_plugin.cpp
+  mor_add_library( toolbox_heat
+    SRCS toolboxmor_heat.cpp
     LINK_LIBRARIES  feelpp_toolbox_heat_lib_2dP1G1 feelpp_toolbox_heat_lib_3dP1G1 Feelpp::feelpp_mor
-    EXEC CRB_LIB_TOOLBOXMOR_PLUGIN_HEAT )
+    EXEC CRB_LIB_TOOLBOX_HEAT )
+  target_compile_definitions(${CRB_LIB_TOOLBOX_HEAT} PRIVATE FEELPP_MOR_PLUGIN_NAME=toolbox_heat FEELPP_MOR_PLUGIN_LIBNAME=${CRB_LIB_TOOLBOX_HEAT})
 
 
 
-  feelpp_add_application(heat_offline SRCS toolboxmor_heat.cpp
+  feelpp_add_application(toolbox_heat_offline SRCS toolboxmor_heat_offline.cpp
     LINK_LIBRARIES
-    ${CRB_LIB_TOOLBOXMOR_PLUGIN_HEAT}
+    ${CRB_LIB_TOOLBOX_HEAT}
     #feelpp_toolbox_heat_lib_2dP1G1
     #feelpp_toolbox_heat_lib_3dP1G1
     #Feelpp::feelpp_mor
@@ -21,7 +22,7 @@ if ( TARGET Feelpp::feelpp_toolbox_heat_lib )
   add_dependencies(feelpp_mor_apps ${MOR_TOOLBOX_HEAT} )
 
 endif()
-if (1)
+if (0)
 if ( TARGET Feelpp::feelpp_toolbox_coefficientformpdes_lib)    
     feelpp_add_application(cfpdes SRCS toolboxmor_cfpdes.cpp LINK_LIBRARIES Feelpp::feelpp_toolbox_coefficientformpdes_lib Feelpp::feelpp_mor INSTALL EXEC MOR_TOOLBOX_CFPDES)
     add_dependencies(feelpp_mor_apps ${MOR_TOOLBOX_CFPDES} )
