+++ mor/apps/thermalfin-2D/thermal-fin.cfg
@@ -5,24 +5,25 @@ journal=false
 
 [toolboxmor]
 filename=$cfgdir/thermal-fin.json
-trainset-deim-size=40
-trainset-mdeim-size=40
+trainset-deim-size=40#1000#40
+trainset-mdeim-size=40#1000#40
 
 [heat]
 filename=$cfgdir/thermal-fin.json
 mesh.filename=$cfgdir/fin2d.geo
-# gmsh.hsize=0.01
+gmsh.hsize=0.05#0.01
 use_velocity-convection=false   
 initial-solution.temperature=300
 verbose_solvertimer=true
 ksp-monitor=true
-pc-type=lu
+pc-type=gamg#lu
 use-cst-vector=false
 use-cst-matrix=false
 
 [crb]
 results-repo-name=thermal-fin-heat
 rebuild-database=true
+#db.update=4 # 4 for new dbid
 output-index=0
 error-type=2
 use-random-WNmu=50
@@ -37,7 +38,7 @@ orthonormality-tol=1e-12#1e-8
 orthonormalize-dual=true
 # absolute-error=true
 # solve-fem-monolithic=true
-solve-dual-problem=false
+solve-dual-problem=true#false
 # use-primal-pc=false
 # use-fast-eim=false
 compute-fem-during-online=true
