+++ mor/apps/toolboxmor_heat.cpp
@@ -1,301 +1,105 @@
-#include <fmt/core.h>
-#include <fmt/compile.h>
+#include "toolboxmor_heat.hpp"
 
-#include "toolboxmor_heat_plugin.hpp"
-#include <feel/feelmor/toolboxmor.hpp>
 #include <feel/feelmodels/heat/heat.hpp>
+#include <feel/feelmor/crbplugin.hpp>
 
-using namespace Feel;
-
-void writeErrors(std::ostream& out, std::vector<std::vector<double> > const& err)
+namespace Feel
 {
-    if( out && Environment::isMasterRank() )
-    {
-        int N = err.size();
-        int size = err[0].size();
-        out << std::setw(5) << "N";
-        for(int i = 0; i < size; ++i)
-            out << std::setw(24) << "mu_" << i;
-        out << std::endl;
-        for(int n = 0; n < N; ++n)
-        {
-            out << std::setw(5) << n+1;
-            for(int i = 0; i < size; ++i)
-                out << std::setw(25) << err[n][i];
-            out << std::endl;
-        }
-    }
+
+template<typename ToolboxType, int Options>
+ToolboxMorHeat<ToolboxType,Options>::ToolboxMorHeat( std::string const& name, std::string const& prefix )
+    :
+    super_type( name, prefix )
+{
+    this->setPluginName( BOOST_PP_STRINGIZE(FEELPP_MOR_PLUGIN_NAME) + fmt::format("_{}dP{}G{}",toolbox_type::nDim, toolbox_type::nOrderTemperature, toolbox_type::nOrderGeo) );
+    this->setPluginLibName( BOOST_PP_STRINGIZE(FEELPP_MOR_PLUGIN_LIBNAME) );
 }
 
-std::tuple<std::vector<double>,std::vector<double>,std::vector<double>,std::vector<double>>
-computeStats(std::vector<std::vector<double>> const& errs)
+template<typename ToolboxType, int Options>
+std::shared_ptr<typename ToolboxMorHeat<ToolboxType,Options>::self_type>
+ToolboxMorHeat<ToolboxType,Options>::createReducedBasisModel()
 {
-    int N = errs.size();
-    std::vector<double> min(N), max(N), mean(N), stdev(N);
-    if( N == 0 )
-        return std::make_tuple(min, max, mean, stdev);
-    int size = errs[0].size();
-    for(int n = 0; n < N; ++n)
-    {
-        min[n] = *std::min_element(errs[n].begin(), errs[n].end());
-        max[n] = *std::max_element(errs[n].begin(), errs[n].end());
-        double s = std::accumulate(errs[n].begin(), errs[n].end(), 0.0);
-        mean[n] = s/size;
-        double accum = std::accumulate(errs[n].begin(), errs[n].end(), 0.0,
-                                       [s,size](double a, double b) {
-                                           return a + (b-s/size)*(b-s/size);
-                                       });
-        stdev[n] = accum/size;
-    }
-    return std::make_tuple(min, max, mean, stdev);
-}
-
-void writeStats(std::ostream& out, std::vector<double> const& min, std::vector<double> const& mean, std::vector<double> const& max, std::vector<double> const& stdev)
-{
-    if( out && Environment::isMasterRank() )
-    {
-        int N = min.size();
-        out << std::setw(5) << "N" << std::setw(25) << "min" << std::setw(25) << "max"
-            << std::setw(25) << "mean" << std::setw(25) << "stdev" << std::endl;
-        for(int n = 0; n < N; ++n)
-            out << std::setw(5) << n+1 << std::setw(25) << min[n] << std::setw(25) << max[n]
-                << std::setw(25) << mean[n] << std::setw(25) << stdev[n] << std::endl;
-    }
-}
-
-std::vector<double> computeOutputs(std::vector<std::vector<std::vector<Eigen::VectorXd>>> const& Lqm_pr, 
-                                   std::vector<std::vector<std::vector<double>>> const& beta,
-                                   Eigen::VectorXd const& uN)
-{
-    std::vector<double> outputs;
-    for(int i = 1; i < beta.size(); ++i )
-    {
-        Eigen::VectorXd F_pr = Eigen::VectorXd::Zero(uN.size());
-        for(int q = 0; q < beta[i].size(); ++q)
-            for(int m = 0; m < beta[i][q].size(); ++m)
-                F_pr += beta[i][q][m]*Lqm_pr[i-1][q][m].head(uN.size());
-        outputs.push_back(F_pr.dot(uN));
-    }
-    return outputs;
-}
-
-template<int Dim, int Order>
-int runSimulation()
-{
-    using convex_type = Simplex<Dim>;
-    using base_type = Lagrange<Order, Scalar, Continuous, PointSetFekete>;
-    using heat_tb_type = FeelModels::Heat<convex_type, base_type>;
-    using heat_tb_ptrtype = std::shared_ptr<heat_tb_type>;
-    using space_type = typename heat_tb_type::space_temperature_type;
-
-    using rb_model_type = ToolboxMor<space_type>;
-    using rb_model_ptrtype = std::shared_ptr<rb_model_type>;
-    using crb_model_type = CRBModel<rb_model_type>;
-    using crb_model_ptrtype = std::shared_ptr<crb_model_type>;
-    using crb_type = CRB<crb_model_type>;
-    using crb_ptrtype = std::shared_ptr<crb_type>;
-    using wn_type = typename crb_type::wn_type;
-    using vectorN_type = Eigen::VectorXd;
-    using export_vector_wn_type = typename crb_type::export_vector_wn_type;
-    using mesh_type = typename rb_model_type::mesh_type;
-    using mesh_ptrtype = typename rb_model_type::mesh_ptrtype;
-    using parameter_type = typename rb_model_type::parameter_type;
-    using sampling_type = typename crb_type::sampling_type;
-    using sampling_ptrtype = std::shared_ptr<sampling_type>;
-
-    using deim_function_type = typename rb_model_type::deim_function_type;
-    using mdeim_function_type = typename rb_model_type::mdeim_function_type;
-#if 0
-    auto heatBox = heat_tb_type::New(_prefix="heat");
+    auto heatBox = toolbox_type::New(_prefix="heat");
     heatBox->init();
     heatBox->printAndSaveInfo();
 
-    rb_model_ptrtype model = std::make_shared<rb_model_type>(soption("toolboxmor.name"));
-    model->setFunctionSpaces(heatBox->spaceTemperature());
-    auto heatBoxModel = DeimMorModelToolbox<heat_tb_type>::New(heatBox);
-    model->initToolbox(heatBoxModel);
-#else
-    using rb_model_heat_type = ToolboxMorPlugin<heat_tb_type>;
-    auto model = rb_model_heat_type::createReducedBasisModel();
-    auto heatBox = model->offlineToolbox();
-#endif
+    auto model = std::make_shared<self_type>(soption("toolboxmor.name"));
+    model->initOffline( heatBox );
 
-    crb_model_ptrtype crbModel = std::make_shared<crb_model_type>(model);
-    crb_ptrtype crb = crb_type::New(soption("toolboxmor.name"), crbModel, crb::stage::offline);
+    return model;
+}
 
-    tic();
-    crb->offline();
-    toc("offline");
+template<typename ToolboxType, int Options>
+void
+ToolboxMorHeat<ToolboxType,Options>::initOffline( toolbox_ptrtype toolbox )
+{
+    M_offlineToolbox = toolbox;
+    this->setFunctionSpaces( toolbox->spaceTemperature() );
+    auto heatBoxModel = DeimMorModelToolbox<toolbox_type>::New( toolbox );
 
-    if( !boption("toolboxmor.do-cvg") )
-        return 0;
+    if ( M_offlineToolbox->hasModelProperties() )
+        this->addModelData( "toolbox_json_setup", M_offlineToolbox->modelProperties().jsonData(), "toolbox_model/setup.json" );
 
-    // convergence study
-    int N = crb->dimension();
-    int timeSteps = 1;
-    std::vector<vectorN_type> uNs(timeSteps, vectorN_type(N)), uNolds(timeSteps, vectorN_type(N));
-    std::vector<double> outs(timeSteps, 0);
+    this->initToolbox(heatBoxModel);
 
-    auto allOutputs = model->modelProperties()->outputs();
-    auto outputs = allOutputs.ofTypes({"integrate","mean","sensor","point"});
-    auto Fqm = model->getFqm();
-    std::vector<std::vector<std::vector<Eigen::VectorXd>>> Lqm_pr(Fqm.size()-1);
-    for( int i = 1; i < Fqm.size(); ++i)
+    this->initOnline();// maybe give here heatBoxModel
+}
+
+template<typename ToolboxType, int Options>
+void
+ToolboxMorHeat<ToolboxType,Options>::initOnline()
+{
+    auto heatBoxModel = DeimMorModelToolbox<toolbox_type>::New("heat");
+
+    auto modelProps = std::make_shared<ModelProperties>();// this->repository().expr(),this->worldCommPtr(), this->prefix(), this->clovm() );
+    if( this->hasModelData("toolbox_json_setup") )
     {
-        Lqm_pr[i-1].resize(Fqm[i].size());
-        for( int q = 0; q < Fqm[i].size(); ++q )
-        {
-            Lqm_pr[i-1][q].resize(Fqm[i][q].size());
-            for( int m = 0; m < Fqm[i][q].size(); ++m )
-            {
-                Lqm_pr[i-1][q][m] = Eigen::VectorXd(N);
-                for( int n = 0; n < N; ++n )
-                    Lqm_pr[i-1][q][m](n) = crbModel->Fqm( i, q, m, crbModel->rBFunctionSpace()->primalBasisElement(n) );
-            }
-        }
+        auto & mdata = this->additionalModelData("toolbox_json_setup");
+        auto const& jsonData = mdata.template fetch_data<nl::json>( this->crbModelDb().dbRepository() );
+        modelProps->setup( jsonData );
     }
 
-    sampling_ptrtype sampling( new sampling_type( crbModel->parameterSpace() ) );
-    int size = ioption("toolboxmor.sampling-size");
-    sampling->clear();
-    sampling->randomize( size, true );
+    heatBoxModel->setToolboxInitFunction(
+        [modelProps](   /*auto*/ typename toolbox_type::mesh_ptrtype  mesh ) {
+            auto tbDeim = std::make_shared<toolbox_type>( "heat"/*M_prefix*/);
+            tbDeim->setModelProperties( modelProps );
+            tbDeim->setMesh(mesh);
+            tbDeim->init();
+            //tbDeim->printAndSaveInfo();
+            return tbDeim;
+        });
 
-    std::vector<std::vector<double> > errs(N, std::vector<double>(size)), errsRel(N, std::vector<double>(size));
-    std::vector<std::vector<std::vector<double>>> errsOutput(Lqm_pr.size(), std::vector<std::vector<double>>(N, std::vector<double>(size)));
-    std::vector<std::vector<std::vector<double>>> errsOutputRel(Lqm_pr.size(), std::vector<std::vector<double>>(N, std::vector<double>(size)));
-    std::vector<std::vector<double>> errsOutputRef(N, std::vector<double>(size)), errsOutputRefRel(N, std::vector<double>(size));
-    std::vector<double> output;
-    auto Xh = model->functionSpace();
-    auto mesh = Xh->mesh();
-    auto rangeT = elements(support(Xh));
-    auto TFE = Xh->element();
-    auto TRB = Xh->element();
+    this->initOnlineToolbox(heatBoxModel);
+}
 
-    Feel::cout << "starting convergence study with " << size << " random parameters" << std::endl;
-    int j = 0;
-    for( auto const& mu : *sampling )
-    {
-        Feel::cout << "cvg for parameter mu=" << mu.toString() << std::endl;
-        for( int i = 0; i < mu.size(); ++i )
-            heatBox->addParameterInModelProperties(mu.parameterName(i), mu(i));
-        heatBox->updateParameterValues();
-        heatBox->solve();
-        TFE = heatBox->fieldTemperature();
-        auto normT = normL2( _range=rangeT, _expr=idv(TFE) );
-        int k = 1;
-        std::vector<double> outputsFE;
-        model->computeBetaQm(mu);
-        for( auto& [name, output] : outputs )
-            outputsFE.push_back(model->output(k++, mu, TFE));
-
-        auto betaFqm = model->computeBetaQm(mu).template get<1>();
-        for(int n = 0; n < N; ++n)
-        {
-            crb->fixedPointPrimal(n+1, mu, uNs, uNolds, outs);
-            vectorN_type uN = uNs[0];
-            TRB = crb->expansion( uN, n+1 );
-            errs[n][j] = normL2( _range=rangeT, _expr=idv(TRB)-idv(TFE) );
-            errsRel[n][j] = errs[n][j]/normT;
-
-            auto outputsRB = computeOutputs(Lqm_pr, betaFqm, uN);
-            for( int k = 0; k < outputsRB.size(); ++k )
-            {
-                errsOutput[k][n][j] = std::abs(outputsRB[k]-outputsFE[k]);
-                errsOutputRel[k][n][j] = errsOutput[k][n][j]/std::abs(outputsFE[k]);
-            }
-        }
-        ++j;
-    }
-
-    fs::ofstream cvgErr( "err.dat" );
-    writeErrors(cvgErr, errs);
-    cvgErr.close();
-    fs::ofstream cvgErrR( "errR.dat" );
-    writeErrors(cvgErrR, errsRel);
-    cvgErrR.close();
-
-    auto [min,max,mean,stdev] = computeStats(errsRel);
-    fs::ofstream cvgStat( "stat.dat" );
-    writeStats(cvgStat, min, mean, max, stdev);
-    cvgStat.close();
-    Feel::cout << "stats on field over size of RB" << std::endl;
-    writeStats(std::cout, min, mean, max, stdev);
-
-    double tol = doption("toolboxmor.tolerance");
-    bool status = mean[N-1] < tol;
-
-    int k = 0;
-    for( auto const& [name,output] : outputs)
-    {
-        fs::ofstream cvgErrO( "err_"+name+".dat" );
-        writeErrors(cvgErrO, errsOutput[k]);
-        cvgErrO.close();
-        fs::ofstream cvgErrOR( "errR_"+name+".dat" );
-        writeErrors(cvgErrOR, errsOutputRel[k]);
-        cvgErrOR.close();
-        auto [minO,maxO,meanO,stdevO] = computeStats(errsOutputRel[k]);
-        fs::ofstream cvgStatO( "stat_"+name+".dat" );
-        writeStats(cvgStatO, minO, meanO, maxO, stdevO);
-        cvgStatO.close();
-        Feel::cout << "stats on output " << name << " over size of RB" << std::endl;
-        writeStats(std::cout, minO, meanO, maxO, stdevO);
-        status = status && meanO[N-1] < tol;
-        k++;
-    }
-
-    auto e = exporter(_mesh=mesh);
-    e->add("TFE", TFE);
-    e->add("TRB", TRB);
-    e->save();
-
-    return !status;
+template<typename ToolboxType, int Options>
+void
+ToolboxMorHeat<ToolboxType,Options>::setupSpecificityModel( boost::property_tree::ptree const& ptree, std::string const& dbDir )
+{
+    super_type::setupSpecificityModel( ptree,dbDir );
+    this->initOnline();
 }
 
 
-int main( int argc, char** argv)
+template <int Dim,int Order>
+struct heat_type
 {
-    using namespace Feel;
-    try
-    {
-        po::options_description opt("options");
-        opt.add_options()
-            ("case.dimension", Feel::po::value<int>()->default_value( 3 ), "dimension")
-            ("case.discretization", Feel::po::value<std::string>()->default_value( "P1" ), "discretization : P1,P2,P3 ")
-            ( "toolboxmor.name", po::value<std::string>()->default_value( "toolboxmor" ), "Name of the db directory" )
-            ( "toolboxmor.do-cvg", po::value<bool>()->default_value( true ), "do convergence test" )
-            ( "toolboxmor.sampling-size", po::value<int>()->default_value(10), "size of the sampling" )
-            ( "toolboxmor.tolerance", po::value<double>()->default_value(5e-2), "tolerance" )
-            ;
+    using convex_type = Simplex<Dim>;
+    using base_type = Lagrange<Order, Scalar, Continuous, PointSetFekete>;
+    using type = FeelModels::Heat<convex_type, base_type>;
+};
 
-        Environment env( _argc=argc, _argv=argv,
-                         _desc=opt.add(makeToolboxMorOptions())
-                         .add(toolboxes_options("heat")) );
+using heat_2dP1G1_t = typename heat_type<2,1>::type;
+using heat_3dP1G1_t = typename heat_type<3,1>::type;
 
-        int dimension = ioption(_name="case.dimension");
-        std::string discretization = soption(_name="case.discretization");
-        
-#if 0
-        int status = 0;
-        hana::for_each( Pc_t<>,
-                        [&discretization, &dimension, &status]( auto const& d ) {
-                            constexpr int _dim = std::decay_t<decltype( hana::at_c<0>( d ) )>::value;
-                            constexpr int _torder = std::decay_t<decltype( hana::at_c<1>( d ) )>::value;
-                            std::string const& _discretization = hana::at_c<2>( d );
-                            if ( dimension == _dim && discretization == _discretization )
-                                status = runSimulation<_dim, _torder>();
-                        } );
-#else
-        int status = 0;
-        if ( dimension == 2 )
-            status = runSimulation<2,1>();
-        // else
-        //     status = runSimulation<3,1>();
-#endif
-        return status;
-    }
-    catch( ... )
-    {
-        handleExceptions();
-    }
-    return EXIT_FAILURE;
+template class ToolboxMorHeat<heat_2dP1G1_t>;
+template class ToolboxMorHeat<heat_3dP1G1_t>;
+
+// FEELPP_CRB_PLUGIN_TEMPLATE( ToolboxMor_heat_2dP1, ToolboxMorHeat<heat_2dP1G1_t>, toolboxmor_heat_2dP1 )
+// FEELPP_CRB_PLUGIN_TEMPLATE( ToolboxMor_heat_3dP1, ToolboxMorHeat<heat_3dP1G1_t>, toolboxmor_heat_3dP1 )
+
+FEELPP_CRB_PLUGIN_TEMPLATE( ToolboxMor_heat_2dP1G1, ToolboxMorHeat<heat_2dP1G1_t>, BOOST_PP_CAT(FEELPP_MOR_PLUGIN_NAME,_2dP1G1) )
+FEELPP_CRB_PLUGIN_TEMPLATE( ToolboxMor_heat_3dP1G1, ToolboxMorHeat<heat_3dP1G1_t>, BOOST_PP_CAT(FEELPP_MOR_PLUGIN_NAME,_3dP1G1) )
+
+
 }
